[traversal] Support additional cluster exit functionality. This makes it more efficient too (sort-by-side one time less). This fixes some cases in buffer when rescaling is turned off.
